memoization methods rename + jsDocs
