+++ src/libs/observable-api/states/deep-state.ts
@@ -2,7 +2,7 @@ import { createObservablePart } from '../utils/create-observable-part.function.j
 import { deepFreeze } from '../utils/deep-freeze.function.js';
 import type { MappingFunction } from '../types/mapping-function.type.js';
 import type { MemoizationFunction } from '../types/memoization-function.type.js';
-import { jsonStringComparison } from '../utils/naive-object-comparison.function.js';
+import { jsonStringComparison } from '../utils/json-string-comparison.function.js';
 import { UmbBasicState } from './basic-state.js';
 
 /**
