+++ src/libs/observable-api/utils/index.ts
@@ -4,7 +4,7 @@ export * from './deep-freeze.function.js';
 export * from './default-memoization.function.js';
 export * from './filter-frozen-array.function.js';
 export * from './merge-observables.function.js';
-export * from './naive-object-comparison.function.js';
+export * from './json-string-comparison.function.js';
 export * from './observe-multiple.function.js';
 export * from './partial-update-frozen-array.function.js';
 export * from './push-to-unique-array.function.js';
