--- src/libs/observable-api/utils/naive-object-comparison.function.ts
@@ -1,3 +1,0 @@
-export function jsonStringComparison(objOne: any, objTwo: any): boolean {
-	return JSON.stringify(objOne) === JSON.stringify(objTwo);
-}
